NP-3515 Minor refactoring and renaming
