+++ src/pages/messages/MyMessagesPage.tsx
@@ -3,15 +3,15 @@ import { PublicationsApiPath } from '../../api/apiPaths';
 import { ListSkeleton } from '../../components/ListSkeleton';
 import { PageHeader } from '../../components/PageHeader';
 import { StyledPageWrapperWithMaxWidth } from '../../components/styled/Wrappers';
-import { SupportRequest } from '../../types/publication_types/messages.types';
+import { PublicationConversation } from '../../types/publication_types/messages.types';
 import { RoleName } from '../../types/user.types';
 import { useFetch } from '../../utils/hooks/useFetch';
 import { MessagesOverview } from '../worklist/MessagesOverview';
 
-const MyMessages = () => {
+const MyMessagesPage = () => {
   const { t } = useTranslation('workLists');
 
-  const [supportRequestsResponse, isLoadingSupportRequests] = useFetch<SupportRequest[]>({
+  const [supportRequestsResponse, isLoadingSupportRequests] = useFetch<PublicationConversation[]>({
     url: `${PublicationsApiPath.Messages}?role=${RoleName.CREATOR}`,
     errorMessage: t('feedback:error.get_messages'),
     withAuthentication: true,
@@ -21,7 +21,7 @@ const MyMessages = () => {
   return (
     <StyledPageWrapperWithMaxWidth>
       <PageHeader>{t('messages')}</PageHeader>
-      {isLoadingSupportRequests && !supportRequestsResponse ? (
+      {isLoadingSupportRequests ? (
         <ListSkeleton minWidth={100} maxWidth={100} height={100} />
       ) : (
         <MessagesOverview conversations={supportRequests} />
@@ -30,4 +30,4 @@ const MyMessages = () => {
   );
 };
 
-export default MyMessages;
+export default MyMessagesPage;
