+++ src/pages/worklist/SupportRequestAccordion.tsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { Link as RouterLink } from 'react-router-dom';
@@ -14,7 +15,6 @@ import { Registration } from '../../types/registration.types';
 import { getRegistrationLandingPagePath } from '../../utils/urlPaths';
 import { MessageList } from './MessageList';
 import { isErrorStatus, isSuccessStatus } from '../../utils/constants';
-import { useState } from 'react';
 import { RootStore } from '../../redux/reducers/rootReducer';
 
 const StyledAccordion = styled(Accordion)`
